Enemy/Kukku receiveMessage: 95.28% -> 96.94% (signed (s32)message<2 compare)
The message<2 test must be signed (cmpwi) not unsigned (cmplwi); message
is u32 in the signature but the original compares it as signed. Residual
is the block-ordering form (bge do; b skip vs blt skip) - currently-hard.

+++ src/Enemy/Kukku.cpp
@@ -81,7 +81,7 @@ BOOL TKukku::receiveMessage(THitActor* sender, u32 message)
 	if (checkLiveFlag(LIVE_FLAG_DEAD))
 		return FALSE;
 
-	if (message < 2 && (s32)message >= 0) {
+	if ((s32)message < 2 && (s32)message >= 0) {
 		mSpine->reset();
 		mSpine->setNext(&TNerveSmallEnemyDie::theNerve());
 		return TRUE;
